pyproject.toml
[tool.poetry]
name = "whisper-api-demo"
version = "0.1.0"
description = ""
authors = ["devyoon91 <kimbyungyoun91@gmail.com>"]
readme = "README.md"
packages = [{include = "whisper_api_demo"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^1.13.1"
openai-whisper = "^20230124"
setuptools-rust = "^1.5.2"
torchvision = "^0.14.1"
torchaudio = "^0.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
